--- src/speculators/models/dflash/utils.py
@@ -1,17 +1,0 @@
-import torch
-from typing import Optional
-from datasets import load_dataset, Features, Sequence, Value
-
-def build_target_layer_ids(num_target_layers: int, num_draft_layers: int):
-    if num_draft_layers == 1:
-        return [(num_target_layers // 2)]
-    start = 1
-    end = num_target_layers - 3
-    span = end - start
-    target_layer_ids = [
-        int(round(start + (i * span) / (num_draft_layers - 1)))
-        for i in range(num_draft_layers)
-    ]
-    return target_layer_ids
-
-
