+++ src/speculators/train/data.py
@@ -9,7 +9,7 @@ from typing import Any
 
 import torch
 import torch.nn.functional as F  # noqa: N812
-from torch.utils.data import Dataset # type: ignore
+from torch.utils.data import Dataset  # type: ignore
 
 from speculators.train.noise_transforms import TransformTensors
