+++ src/speculators/train/utils.py
@@ -3,18 +3,7 @@ import os
 
 import torch
 import torch.distributed as dist
-
-# Handle different PyTorch FSDP APIs across versions
-try:
-    from torch.distributed.fsdp import MixedPrecisionPolicy, fully_shard
-except ImportError:
-    try:
-        # Older PyTorch versions use MixedPrecision instead of MixedPrecisionPolicy
-        from torch.distributed.fsdp import MixedPrecision as MixedPrecisionPolicy, fully_shard
-    except ImportError:
-        # Even older versions - provide no-op implementations
-        MixedPrecisionPolicy = None
-        fully_shard = None
+from torch.distributed.fsdp import MixedPrecisionPolicy, fully_shard
 
 local_rank = int(os.environ.get("LOCAL_RANK", "0"))
 world_size = int(os.environ.get("WORLD_SIZE", "1"))
